refactor(fbs-common): verplaats Identificatienummer + types vanuit berichtenmagazijn
Hergebruikbaar in sessiecache (volgt). Voegt toCanonicalString() toe voor
unieke cache-keys en log-correlatie. Imports in magazijn-module mechanisch
gepatcht; gedragsneutraal.
