+++ services/berichtenmagazijn/src/main/kotlin/nl/rijksoverheid/moz/fbs/berichtenmagazijn/opslag/Bericht.kt
@@ -1,6 +1,8 @@
 package nl.rijksoverheid.moz.fbs.berichtenmagazijn.opslag
 
 import nl.rijksoverheid.moz.fbs.common.exception.requireValid
+import nl.rijksoverheid.moz.fbs.common.identificatie.Identificatienummer
+import nl.rijksoverheid.moz.fbs.common.identificatie.Oin
 import java.time.Instant
 import java.util.UUID
