+++ services/berichtenmagazijn/src/test/kotlin/nl/rijksoverheid/moz/fbs/berichtenmagazijn/opslag/BerichtRetentieIntegrationTest.kt.tmp
@@ -1,0 +1,2 @@
+package nl.rijksoverheid.moz.fbs.berichtenmagazijn.opslag
+
